requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==6.2.0
qrcode==7.3.1
pillow==9.1.1
django-rest-registration==0.7.2
celery==5.2.7
redis==4.3.3
